+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Make domain regex matches even more strict: limit amount of subdomains
 * Automatically remove problematic domains from Greenlock (though it does not work)
 * Keep track of IP reputations
+* Correctly remove site instances from the `dispatcher.domains` object
 
 ## 0.5.3 (2023-10-19)